wip(make_equidistant): introduce deprecation warning for _misc.make_equidistant_
